rails-like router
