+++ app.rb
@@ -1,21 +1,6 @@
-require_relative './router'
+require_relative './config/routes'
 
 class App
-  attr_reader :router
-
-  def initialize
-    @router = Router.new
-
-    # shows the home page
-    router.get('/') { "Akshay's Blog" }
-    
-    # shows all articles
-    router.get('/articles') { 'All Articles' }
-    
-    # shows a single article
-    router.get('/articles/1') { "First Article" }
-  end
-
   def call(env)
     headers = { 'Content-Type' => 'text/html' }
     
@@ -23,4 +8,9 @@ class App
     
     [200, headers, [response_html]]
   end
+
+  private
+    def router
+      Router.instance
+    end
 end
