+++ config/router.rb
@@ -1,0 +1,26 @@
+require 'singleton'
+
+class Router
+  include Singleton
+
+  attr_reader :routes
+
+  class << self
+    def draw(&blk)
+      Router.instance.instance_exec(&blk)
+    end
+  end
+
+  def initialize
+    @routes = Hash.new
+  end
+
+  def get(path, &blk)
+    @routes[path] = blk
+  end
+
+  def build_response(path)
+    handler = @routes[path] || ->{ "no route found for #{path}" }
+    handler.call
+  end
+end
